refactor:
- Changed name of script so clear it contains functions

- Created pytest for this also but not sure if needed
